dogs/requirements.txt
Django==4.2.2
djangorestframework==3.15.2
flake8==5.0.4
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==3.0.0
